backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
asyncpg==0.31.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.14.2
bidict==0.23.1
billiard==4.2.4
bitarray==3.8.0
blinker==1.9.0
boto3==1.42.68
botocore==1.42.68
build==1.4.0
cachetools==6.2.6
celery==5.6.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.5.0
ckzg==2.1.7
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
coverage==7.13.4
cryptography==46.0.5
cytoolz==1.1.0
databases==0.9.0
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.2.0
Django==6.0.3
django-bootstrap5==24.2
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-phonenumber-field==8.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
factory_boy==3.3.3
Faker==40.11.0
fastapi==0.134.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
filelock==3.24.2
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
frozenlist==1.7.0
fsspec==2026.2.0
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.37.1
grpcio==1.78.0
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
imutils==0.5.4
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.19
langchain-groq==1.1.2
langchain-openai==1.1.11
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nassl==5.4.0
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.24.1
openai==2.28.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
passlib==1.7.4
phonenumberslite==8.13.45
pillow==12.1.1
pluggy==1.6.0
postgrest==2.28.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.10.1
PyMuPDF==1.27.1
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyroaring==1.0.3
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.12.2
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socketio==5.13.0
pytz==2024.1
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
qrcode==8.2
razorpay==2.0.1
realtime==2.28.0
redis==6.4.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
rlp==4.1.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.5
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
selenium==4.36.0
sendgrid==6.12.5
sentence-transformers==5.2.3
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.2.0
sslyze==6.3.0
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
structlog==25.5.0
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tls_parser==2.0.2
tokenizers==0.22.2
toolz==1.1.0
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
trio==0.31.0
trio-websocket==0.12.2
twilio==9.2.3
typer==0.24.0
typer-slim==0.24.0
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.37.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
web3==7.14.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
